+++ src/main/java/net/sodiumstudio/befriendmobs/events/EntityEvents.java
@@ -44,6 +44,7 @@ import net.sodiumstudio.befriendmobs.entity.befriending.BefriendableMobInteracti
 import net.sodiumstudio.befriendmobs.entity.befriending.BefriendingHandler;
 import net.sodiumstudio.befriendmobs.entity.befriending.registry.BefriendingTypeRegistry;
 import net.sodiumstudio.befriendmobs.entity.capability.CAttributeMonitor;
+import net.sodiumstudio.befriendmobs.entity.capability.CLivingEntityDelayActionHandler;
 import net.sodiumstudio.befriendmobs.inventory.BefriendedInventory;
 import net.sodiumstudio.befriendmobs.item.MobRespawnerItem;
 import net.sodiumstudio.befriendmobs.item.MobRespawnerInstance;
@@ -446,15 +447,11 @@ public class EntityEvents
 		if (!event.getEntity().level.isClientSide)
 		{
 			// Tick attribute monitor
-			event.getEntity().getCapability(BMCaps.CAP_ATTRIBUTE_MONITOR).ifPresent((cap) -> 
-			{
-				cap.tick();
-			});
+			event.getEntity().getCapability(BMCaps.CAP_ATTRIBUTE_MONITOR).ifPresent(CAttributeMonitor::tick);
 			// Tick item stack monitor
-			event.getEntity().getCapability(BMCaps.CAP_ITEM_STACK_MONITOR).ifPresent((cap) ->
-			{
-				cap.tick();
-			});
+			event.getEntity().getCapability(BMCaps.CAP_ITEM_STACK_MONITOR).ifPresent(CItemStackMonitor::tick);
+			// Tick delay action handler
+			event.getEntity().getCapability(BMCaps.CAP_DELAY_ACTION_HANDLER).ifPresent(CLivingEntityDelayActionHandler::tick);
 			if (event.getEntity() instanceof Mob mob)
 			{
 				// update befriendable mobs