+++ src/main/java/net/sodiumstudio/befriendmobs/registry/BMCapabilityAttachment.java
@@ -21,9 +21,13 @@ import net.sodiumstudio.befriendmobs.entity.capability.CAttributeMonitorProvider
 import net.sodiumstudio.befriendmobs.entity.capability.CBMPlayerModule;
 import net.sodiumstudio.befriendmobs.entity.capability.CBefriendableMobProvider;
 import net.sodiumstudio.befriendmobs.entity.capability.CHealingHandlerProvider;
+import net.sodiumstudio.befriendmobs.entity.capability.CLivingEntityDelayActionHandler;
+import net.sodiumstudio.befriendmobs.entity.capability.wrapper.IAttributeMonitor;
+import net.sodiumstudio.befriendmobs.entity.capability.wrapper.ILivingDelayActions;
 import net.sodiumstudio.befriendmobs.item.baublesystem.CBaubleDataCache;
 import net.sodiumstudio.befriendmobs.item.baublesystem.IBaubleEquipable;
 import net.sodiumstudio.befriendmobs.item.capability.CItemStackMonitor;
+import net.sodiumstudio.befriendmobs.item.capability.wrapper.IItemStackMonitor;
 import net.sodiumstudio.befriendmobs.level.CBMLevelModule;
 
 @Mod.EventBusSubscriber(bus = Mod.EventBusSubscriber.Bus.FORGE)
@@ -37,12 +41,24 @@ public class BMCapabilityAttachment {
 		if (event.getObject() instanceof LivingEntity living)
 		{
 			// Attribute change monitor
-			CAttributeMonitorProvider prvd = new CAttributeMonitorProvider(living);
-			event.addCapability(new ResourceLocation(BefriendMobs.MOD_ID, "cap_attribute_monitor")
-					, prvd);		
+			if (living instanceof IAttributeMonitor)
+			{
+				CAttributeMonitorProvider prvd = new CAttributeMonitorProvider(living);
+				event.addCapability(new ResourceLocation(BefriendMobs.MOD_ID, "cap_attribute_monitor")
+						, prvd);
+			}
 			// Item Stack monitor
-			CItemStackMonitor.Prvd prvd1 = new CItemStackMonitor.Prvd(living);
-			event.addCapability(new ResourceLocation(BefriendMobs.MOD_ID, "cap_item_stack_monitor"), prvd1);
+			if (living instanceof IItemStackMonitor)
+			{
+				CItemStackMonitor.Prvd prvd1 = new CItemStackMonitor.Prvd(living);
+				event.addCapability(new ResourceLocation(BefriendMobs.MOD_ID, "cap_item_stack_monitor"), prvd1);
+			}
+			// Delay action handler
+			if (living instanceof ILivingDelayActions)
+			{
+				CLivingEntityDelayActionHandler.Prvd prvd = new CLivingEntityDelayActionHandler.Prvd(living);
+				event.addCapability(new ResourceLocation(BefriendMobs.MOD_ID, "cap_delay_action_handler"), prvd);
+			}
 		}