+++ src/main/java/net/sodiumstudio/nautils/math/GeometryUtil.java
@@ -1,6 +1,12 @@
 package net.sodiumstudio.nautils.math;
 
+import java.util.Random;
+
+import com.mojang.logging.LogUtils;
+
 import net.minecraft.world.phys.AABB;
+import net.minecraft.world.phys.Vec2;
+import net.minecraft.world.phys.Vec3;
 
 /**
  * Mathematical utilities related to 3D geometry e.g. Vec3, AABB, etc.
@@ -8,6 +14,8 @@ import net.minecraft.world.phys.AABB;
 public class GeometryUtil
 {
 
+	public static Random rnd = new Random();
+	
 	/**
 	 * Get the squared closest distance of the surfaces of two boxes.
 	 */
@@ -44,5 +52,64 @@ public class GeometryUtil
 			dist = Math.min(dist, getBoxSurfaceDistSqr(box, irregularShape[i]));
 		}
 		return dist;
+	}
+	
+	/**
+	 * Get the angle (theta) from vector xy, in radians. Range: (-pi, pi]
+	 */
+	public static double getAngleRadian(double vec2X, double vec2Y)
+	{
+		Vec3 v = new Vec3(vec2X, vec2Y, 0).normalize();
+		if (vec2X * vec2X + vec2Y * vec2Y < 1e-12d)
+		{
+			LogUtils.getLogger().warn("NaUtils: GeometryUtils#getAngleRadian: input vector is too short. Result may be inaccurate.");
+		}
+		// Handle zero cases
+		if (Math.abs(v.x) == 0d && Math.abs(v.y) == 0d)
+			return 0;
+		else if (Math.abs(v.x) == 0d)
+			return (v.y > 0 ? Math.PI / 2 : -Math.PI / 2);
+		else if (Math.abs(v.y) == 0d)
+			return (v.x > 0 ? 0 : Math.PI);
+		// 1st & 2nd quadrant
+		if (v.y > 0)
+			return Math.acos(v.x);
+		// 4th quadrant
+		else if (v.x > 0)
+			return Math.asin(v.y);
+		// 3rd quadrant
+		else return Math.acos(-v.x) - Math.PI;
+	}
+	
+	
+	/**
+	 * Rotate a Vec3 around Y axis with given angle in degrees. Positive angle represents anticlockwise rotation looked from upside. 
+	 */
+	public static Vec3 rotateY(Vec3 v, double angleDegrees)
+	{
+		double xzScale = Math.sqrt(v.x * v.x + v.y * v.y + v.z * v.z);
+		if (xzScale < 1e-12d)
+		{
+			LogUtils.getLogger().warn("NaUtils: GeometryUtils#rotateY: input vector is too short. Result may be inaccurate.");
+			return v;
+		}
+		Vec3 xzNorm = new Vec3(v.x, 0d, v.z).normalize();
+		double angleRadians = angleDegrees * Math.PI / 180d;
+		double initAngleRadian = getAngleRadian(xzNorm.x, xzNorm.z);
+		return new Vec3(Math.cos(initAngleRadian + angleRadians), 0d, Math.sin(initAngleRadian + angleRadians))
+				.scale(xzScale).add(0, v.y, 0);
+	}
+	
+	/**
+	 * Get a random unit vector with uniform-distribution on sphere surface area.
+	 */
+	public static Vec3 randomVector()
+	{
+		while (true)
+		{
+			Vec3 v = new Vec3(rnd.nextDouble() * 2d - 1d, rnd.nextDouble() * 2d - 1d, rnd.nextDouble() * 2d - 1d);
+			if (v.x * v.x + v.y * v.y + v.z * v.z <= 1d)
+				return v.normalize();
+		}
 	}
 }